Remove a trailing \
Remove a trailing \ from a line. The character is not strictly necessary
and it currently differs between sc-dev and master, which makes the
removal of the line in sc-dev conflict in the automerger. It's easiest
to fix the line to be the same in master than it is to cherry-pick /
land the topic in master.

Bug: 178169250
Test: treehugger only
Change-Id: I9b5c743d8744c6e9382bcecacfed76f050539eda