Move copy rule for system server jars from dexpreopt_gen to makefiles.
This is necessary in order to expose the copy rule to Ninja. Otherwise
Ninja has a build rule that depends on the copied jar (namely, dexpreopt
commands for all subsequent system server jars that have the current one
in class loader context), but no rule that creates it, so Ninja fails
with a "no rule to make ..." error.

The change only affects system server ed in Android.mk

Previously the problem existed, but was hidden by the fact that the only
system server jar defined in Android.mk was the last one on the list, so
no other jar depended on it. Now that updatable apex jars are also
dexpreopted (https://r.android.com/1828115) the problem was uncovered.

Also the patch removes obsolete logic that disable dexpreopt for
PRODUCT_APEX_SYSTEM_SERVER_JARS; these jars are now preopted (but they
are all defined in Android.bp anyway, so no functional change here).

Bug: 200297762
Test: lunch bertha_x86_64-userdebug && m
Change-Id: I67c6d69f45ca3495f62994b5329f9e424dda4e65