Add Soong variable for optimizing SystemUI Java
Introduce a Soong variable to allow selective enabling of Java
optimizations for SystemUI. This is analogous to the
`SYSTEM_OPTIMIZE_JAVA` flag, but scoped just to SystemUI,
decoupling the feature from system server-related development.

Opting in to optimizations can be achieved with either:
  * Env:
export SYSTEMUI_OPTIMIZE_JAVA=true
  * Make:
$(call add_soong_config_var_value,ANDROID,SYSTEMUI_OPTIMIZE_JAVA,true)

Note that the build rule changes to hook into this variable will
land separately.

Bug: 203088572
Test: `SYSTEMUI_OPTIMIZE_JAVA=true m` (observe env variable change log)
Change-Id: I24b4b76cdf844e3c00735e19e69898867162c81d