Prevent enabling prebuilts by default for builds that don't support them
Except for santizer and framework Java coverage tests as they require
specialized prebuilts built from a matching configuration and we do not
provide them.

And products that use dex2oat(d)s as they are not available as a
prebuilt.

And products that cross compile to arm64 and linux_bionic on the host
as prebuilts for that are not available.

And atv/tv products as they break when building against prebuilts.

(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:d9f5861751e19aee33bad5cd12fa1ae5fb83cc98)

This uses a completely separate Change-Id as the previous change was
applied to lots of downstream branches (so they contain the Change-Id
as far as the automerger is concernd) but the change got discarded due
to some merge issues. Using the previous Change-Id would cause
confusion. The title of this change was also updated as the previous
title only made sense within the larger topic to which it belonged.

Bug: 230478310
Bug: 223592962
Test: presubmit
Change-Id: Ie35ff34f4626070d1a3fe0bb3a6a999a596d0078