Allow non-compat suite to locate testcases dir properly
cc_test that can run on hostside are default to be part of
host-unit-tests suite. The suite is set up similar to device-tests and
general-tests, and it's not a compatibility suite. So following variable
will always be empty:
$(COMPATIBILITY_TESTCASES_OUT_host-unit-tests.
It should use HOST_OUT_TESTCASES as testcases dir.

The same logic should be applied to art-host-tests.

Bug: None
Test: build
Change-Id: Ieddacd5157da73c6df473c103be109d9c3a55281