Target dist files not dist goals.
It turns out dist goals do not map well to units of code distributed.
They often build a combination of files to distribute and files for
internal use only.

This change creates license texts and lists of projects to share for
individual files in dist instead of for dist goals.

If a distribution needs the license texts for part of a dist file or
for a combination of multiple files, whoever is distributing will have
to create a suitable text notice using `gen_notice{}` or `testnotice`
and include it among the dist files for the goal using `dist-for-goals`
or equivalent.

Bug: 151177513
Bug: 210912771
Bug: 235333302
Bug: 232138105
Bug: 235302632

Test: m cts dist

Test: m droid dist

Test: tapas google-tradefed-all && m dist

Change-Id: Id78f032575a2ac1a3cc70643b3ae4d2e6c32da04