Revert "Revert "Target dist files not dist goals.""
This reverts commit c6b02e6fb847384f1a8415b29bd6d78d7dce1543.

Reason for revert: Fixed latent bug in OTA test

This change creates license texts and lists of projects to share for
individual files in dist instead of for dist goals.

If a distribution needs the license texts for part of a dist file or
for a combination of multiple files, whoever is distributing will have
to create a suitable text notice using `gen_notice{}` or `testnotice`
and include it among the dist files for the goal using `dist-for-goals`
or equivalent.

Bug: 151177513
Bug: 210912771
Bug: 235333302
Bug: 232138105
Bug: 235302632

Test: m cts dist

Test: m droid dist

Test: tapas google-tradefed-all && m dist

Change-Id: I9fbcb07acfa7b14a140563cb031374160b6b23c2
Merged-in: I9fbcb07acfa7b14a140563cb031374160b6b23c2