Target dist files not dist goals.
This change creates license texts and lists of projects to share for
individual files in dist instead of for dist goals.

If a distribution needs the license texts for part of a dist file or
for a combination of multiple files, whoever is distributing will have
to create a suitable text notice using `gen_notice{}` or `testnotice`
and include it among the dist files for the goal using `dist-for-goals`
or equivalent.

Bug: 240342946

Test: m cts dist

Test: m droid dist

Test: tapas google-tradefed-all && m dist

Change-Id: Iaea79fc1ecf34e2ace6403d8ef86c667dc9937b2
Merged-in: Iaea79fc1ecf34e2ace6403d8ef86c667dc9937b2
Merged-in: I9fbcb07acfa7b14a140563cb031374160b6b23c2