Update minimum supported targetSdk to 28
This change will cause a dismissable warning message to display
indicating the app is targeting a low targetSdkVersion.

Previously this message was displayed for apps targeting < 23.

The eventual goal is to move this to 29, but still
working through some failed CTS tests. Updating first
to 28, and then will follow up with an update to 29.

Bringing this to 29 would be in line with trailing
by 4 major releases.

Test: make
Test: atest CtsWindowManagerSdk25TestCases:AspectRatioSdk25Tests
Test: atest CtsWindowManagerSdk28TestCases:AspectRatioSdk28Tests
Change-Id: I669d325770b5abebcb29c2f4dc31f82fb8e638c7