Make com.android.sepolicy mandatory
The apex_sepolicy files were removed from the system partition in
ag/17753690 because they were not necessary in T builds.

For updatable sepolicy in Android U+ we are planning to move parts of
the existing platform policy into the apex sepolicy, so we will need to
make the apex_sepolicy files available to all builds.

Bug: 243923977
Test: adb shell ls /dev/selinux /system/etc/selinux
Change-Id: I24b8b8880149885d98caf992afdae908ead11fcf