Disable riscv64 dependency on bcc and ld.mc
riscv64 can't build the device version of bcc and ld.mc due to a
dependency on an old version of LLVM, but they are listed in
base_system.mk which can't add them conditionally based on the target
architecture.  Add a hack to base_system.mk to silence the warning
that the host-only modules are listed in PRODUCT_PACKAGES.  This
can be removed once renderscript is removed from the platform.

Test: lunch aosp_riscv64-userdebug && m
Change-Id: I8efd769a34b110dfb838f7016e029369cb9a0f01