Rename rbcrun/cmd to make `go install` work properly
The standard go tooling assumes that a package's binary name should be
the name of the containing folder. Before this change, `go install` or
`go build` (without `-o`) names this binary `cmd`.

Also run `go mod tidy` while we're here, so the tooling doesn't try to
download github.com/nbutton23/zxcvbn-go, which we don't have a local
copy of.

Test: treehugger
Change-Id: I8f79724e928720caf961cdad128655ef31c3e9b2