Make android_certificate_directory public
This should not be necessary because bp2build should emit a package()
statement with default_visibility = public. However, there appears
to be a problem with bp2build's symlink forest, where it doesn't
handle symlinks correctly.

build/target is a symlink to build/make/target, but in the symlink
forest, build/target is a symlink to the source dir's build/target
instead of to the symlink forest's build/make/target. This means
that when reading this build file via the build/target symlink,
you won't get the bp2build-provided default visibility. (or any other
bp2build generated targets)

We should send a proper fix for this, but in the meantime submit this
quick fix to resolve some failing builds.

Bug: 269577299
Test: lunch bertha_x86_64-userdebug, m nothing
Change-Id: Ia45c7f4a0655b1364b735d6eccf25cad70cad134